@@ -4,15 +4,18 @@
   "private": true,
   "homepage": "https://lukehowlett97.github.io/luke-howlett-cv",
   "dependencies": {
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-scripts": "4.0.3",
-    "react-router-dom": "^5.2.0"
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-router-dom": "^5.2.0",
+    "react-scripts": "^5.0.1"
   },
   "devDependencies": {
+    "autoprefixer": "^10.4.19",
     "cross-env": "^7.0.3",
     "gh-pages": "^6.1.1",
-    "rimraf": "^6.0.1"
+    "postcss": "^8.4.40",
+    "rimraf": "^6.0.1",
+    "tailwindcss": "^3.4.7"
   },
   "scripts": {
     "start": "react-scripts start",
